+++ processor/src/test/resources/expected/com/example/repository/RepositoryWithImplicitIdRepository.java
@@ -25,7 +25,8 @@ public abstract class RepositoryWithImplicitIdRepository extends AbstractReposit
       name = "create_packageName_name"
   )
   @Nonnull
-  RepositoryWithImplicitId create(@Nonnull final String packageName, @Nonnull final String name) {
+  public RepositoryWithImplicitId create(@Nonnull final String packageName,
+      @Nonnull final String name) {
     final Arez_RepositoryWithImplicitId entity = new Arez_RepositoryWithImplicitId(packageName,name);
     registerEntity( entity );
     return entity;